Finish up profile styling
