+++ react-front-end/package.json
@@ -16,6 +16,7 @@
     "node-sass": "^4.14.1",
     "query-string": "^7.0.0",
     "react": "^16.14.0",
+    "react-awesome-slider": "^4.1.0",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.2",
     "react-emoji": "^0.5.0",
