--- react-front-end/src/components/Profile/ProfilePic.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const ProfilePic = ({ profileImage }) =>(
-    <div>
-      <img src={profileImage}/>
-    </div>
-  )
-
-export default ProfilePic;
