--- react-front-end/src/components/Profile/UserDetails.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const UserDetails = ({ userName, userCity, userBio }) =>(
-    <div>
-      {userName}
-      {userCity}
-      {userBio}
-    </div>
-  )
-
-export default UserDetails;
